# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Market service API",
        "title": "Avito.Kitchen API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/cart": {
            "get": {
                "tags": [
                    "cart"
                ],
                "summary": "Get cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Cart"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/cart/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Add product to cart",
                "parameters": [
                    {
                        "description": "Add something to cart",
                        "name": "addRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.AddToCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/cart/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Delete product from cart",
                "parameters": [
                    {
                        "description": "Delete product from cart",
                        "name": "deleteRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.DeleteFromCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/cart/order": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Order cart",
                "parameters": [
                    {
                        "description": "Submit an order for the cart",
                        "name": "orderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.OrderReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/filter": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "market"
                ],
                "summary": "Filter products",
                "parameters": [
                    {
                        "description": "Filter criteria",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Filter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/markets": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Get all markets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Market"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/product": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Set product",
                "parameters": [
                    {
                        "description": "Set product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "market"
                ],
                "summary": "Get product card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "description": "Delete product by ID",
                        "name": "deleteRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.DeleteProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Cart": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "in_process": {
                    "type": "boolean"
                },
                "ordered_at": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Filter": {
            "type": "object",
            "required": [
                "limit"
            ],
            "properties": {
                "limit": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "market_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_price": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 0
                },
                "min_price": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 1
                },
                "offset": {
                    "type": "integer",
                    "minimum": 0
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Market": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "required": [
                "id",
                "market_id",
                "price",
                "title"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "market_id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "maximum": 99999999,
                    "minimum": 1
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "server.AddToCartReq": {
            "type": "object",
            "required": [
                "cart_id",
                "product_id",
                "user_id"
            ],
            "properties": {
                "cart_id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "server.DeleteFromCartReq": {
            "type": "object",
            "required": [
                "cart_id",
                "product_id",
                "user_id"
            ],
            "properties": {
                "cart_id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "server.DeleteProductRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "server.OrderReq": {
            "type": "object",
            "required": [
                "cart_id",
                "user_id"
            ],
            "properties": {
                "cart_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}